pyproject.toml
[tool.poetry]
name = "getlinked_rcb"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sparky-utils = "^1.1.3"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.6.0"
djangorestframework-simplejwt = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
